Default functions in Ingress Manager
